Created Pause Menu and Pause Button
Created Pause Menu and Pause button for both levels.  I added the background for both levels as well.  The Exit button on the top right is the pause button and it works as well as the resume button as a function.  Only issue is on level 1 the items overlap the pause menu when on pause.